% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{ball}
\alias{ball}
\title{Ball Data Set}
\usage{
ball(n = 1000, rad = 1, ndim = 3)
}
\arguments{
\item{n}{Number of points to create.}

\item{rad}{Radius of the ball.}

\item{ndim}{Dimension of the ball.}
}
\value{
Data frame with \code{ndim} coordinate columns and \code{color} containing
RGB colors. For \code{ndim = 3}, coordinates are named \code{x}, \code{y}, \code{z};
otherwise they are named \code{X1}, \code{X2}, ..., \code{Xndim}.
}
\description{
Simulation data randomly sampled from the entire volume of a ball of the
specified dimension (by default, 3-dimensional).
}
\details{
Creates a series of points sampled from an \code{ndim}-dimensional spherical
volume. Points are colored based on the square of their distance from the
origin.
}
\references{
A dataset like this was used in:
Lee, J. A., Peluffo-Ordo'nez, D. H., & Verleysen, M. (2015).
Multi-scale similarities in stochastic neighbour embedding: Reducing
dimensionality while preserving both local and global structure.
\emph{Neurocomputing}, \emph{169}, 246-261.
}
