% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misread_tsne.R
\name{circle_data}
\alias{circle_data}
\title{Uniform 2D Circle}
\usage{
circle_data(n)
}
\arguments{
\item{n}{Number of points per side.}
}
\value{
Data frame with x and y coordinates in the columns \code{x} and
 \code{y} respectively, and color in the \code{color} column.
}
\description{
Circle simulation data function from "How to Use t-SNE Effectively".
}
\details{
Creates a dataset where the points are located on the circumference of a
circle of unit radius, where the angle subtended by the point at the center
of the circle has been uniformly sampled. Each point has a color linearly
mapped from the angle to a rainbow color scheme.
}
\examples{
df <- circle_data(n = 50)
}
\references{
\url{http://distill.pub/2016/misread-tsne/}
}
\seealso{
Other distill.functions: \code{\link{cube_data}},
  \code{\link{gaussian_data}}, \code{\link{grid_data}},
  \code{\link{link_data}}, \code{\link{long_cluster_data}},
  \code{\link{long_gaussian_data}},
  \code{\link{ortho_curve}},
  \code{\link{random_circle_cluster_data}},
  \code{\link{random_circle_data}},
  \code{\link{random_jump}}, \code{\link{random_walk}},
  \code{\link{simplex_data}},
  \code{\link{subset_clusters_data}},
  \code{\link{three_clusters_data}},
  \code{\link{trefoil_data}},
  \code{\link{two_clusters_data}},
  \code{\link{two_different_clusters_data}},
  \code{\link{unlink_data}}
}

