% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misread_tsne.R
\name{cube_data}
\alias{cube_data}
\title{Cube Data}
\usage{
cube_data(n, dim, color = "#003399")
}
\arguments{
\item{n}{Number of points to sample.}

\item{dim}{Dimension of the cube.}

\item{color}{Color to apply to each point.}
}
\value{
Data frame with coordinates in the \code{X1}, \code{X2} ...
 \code{Xdim} columns, and color in the \code{color} column.
}
\description{
Random points sampled from a cube, from "How to Use t-SNE Effectively".
}
\details{
Create a dataset of the of points randomly sampled from a unit cube
of the specified dimension.
}
\examples{
df <- cube_data(n = 50, dim = 3)
}
\references{
\url{http://distill.pub/2016/misread-tsne/}
}
\seealso{
Other distill functions: \code{\link{circle_data}},
  \code{\link{gaussian_data}}, \code{\link{grid_data}},
  \code{\link{link_data}}, \code{\link{long_cluster_data}},
  \code{\link{long_gaussian_data}},
  \code{\link{ortho_curve}},
  \code{\link{random_circle_cluster_data}},
  \code{\link{random_circle_data}},
  \code{\link{random_jump}}, \code{\link{random_walk}},
  \code{\link{simplex_data}},
  \code{\link{subset_clusters_data}},
  \code{\link{three_clusters_data}},
  \code{\link{trefoil_data}},
  \code{\link{two_clusters_data}},
  \code{\link{two_different_clusters_data}},
  \code{\link{unlink_data}}
}
\concept{distill functions}
