% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{curve2d}
\alias{curve2d}
\title{2D Curve Dataset.}
\usage{
curve2d()
}
\value{
Data frame with 1450 rows, and 3 columns: \code{x}, \code{y} columns
  contain the coordinates of the points and \code{color} the RGB color.
}
\description{
Simulation data of a 2D polynomial curve.
}
\details{
This data set is used to assess the behavior of the PaCMAP method of Wang and
co-workers (2021) and some related dimensionality reduction methods,
specifically the effect of initialization on what should be an "easy" dataset
to embed.
 
Points are colored based on their distance along the curve.
}
\references{
Wang, Y., Huang, H., Rudin, C., & Shaposhnik, Y. (2021). 
Understanding how dimension reduction tools work: an empirical approach to 
deciphering t-SNE, UMAP, TriMAP, and PaCMAP for data visualization. 
\emph{J Mach. Learn. Res}, \emph{22}, 1-73.
}
\seealso{
the \href{https://github.com/YingfanWang/PaCMAP}{PaCMAP homepage}.
}
