% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frey.R
\name{frey_faces}
\alias{frey_faces}
\title{Frey Faces dataset}
\format{
A data frame with 1965 rows and 561 variables.
}
\usage{
frey_faces()
}
\value{
The Frey Faces dataset as a dataframe.
}
\description{
Converts Frey Faces Image dataset.
}
\details{
Returns the Frey Faces dataset in a data frame where each row is one image.
This is a series of 1965 images (with dimension 20 x 28) of Brendan Frey's
face taken from sequential frames of a video.

The variables are as follows:
\itemize{
\item \code{px1}, \code{px2}, \code{px3} ... \code{px560} 8-bit grayscale
pixel values (0-255). The pixel index starts at the top right of the image
(\code{px1}) and are then stored row-wise.
\item \code{color} A string representing a color in hex format. It can be
used directly with e.g. the \code{col} parameter in the
\code{\link[graphics]{plot}} function. The color goes from hsl(0, 50, 50)
(red) at frame 1 to hsl(300, 50, 50) (purple) at frame 1965 and on a muted
rainbow scale.
}
}
\note{
Requires the 
\href{https://cran.r-project.org/package=RnavGraphImageData}{RnavGraphImageData}
package to be installed and loaded.
}
\examples{
\dontrun{
frey <- frey_faces()
# PCA Scores plot, with color indicating the frame index
frey_pca <- prcomp(frey[, -561], retx = TRUE, rank. = 2)
plot(frey_pca$x, col = frey$color, pch = 16, cex = 0.75)
}
}
\seealso{
\itemize{
\item{Sam Roweis' dataset web page: \url{http://www.cs.nyu.edu/~roweis/data.html}.}
\item{Each row can be visualized as an image using \code{\link{show_frey_face}}.}
}
}
