% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misread-tsne.R
\name{gaussian_data}
\alias{gaussian_data}
\title{Symmetric Gaussian}
\usage{
gaussian_data(n, dim, sd = 1, color = NULL)
}
\arguments{
\item{n}{Number of points.}

\item{dim}{Dimension of the gaussian.}

\item{sd}{Standard deviation.}

\item{color}{Color to assign to the data points.}
}
\value{
Data frame with coordinates in the \code{X1}, \code{X2} ...
\code{Xdim} columns, and color in the \code{color} column.
}
\description{
Symmetric gaussian data from "How to Use t-SNE Effectively".
}
\details{
Creates a randomly-sampled symmetric gaussian dataset of the specified
dimension.
}
\examples{
df <- gaussian_data(n = 50, dim = 2)
}
\references{
\url{http://distill.pub/2016/misread-tsne/}
}
\seealso{
Other distill functions:
\code{\link[=circle_data]{circle_data()}},
\code{\link[=cube_data]{cube_data()}},
\code{\link[=grid_data]{grid_data()}},
\code{\link[=link_data]{link_data()}},
\code{\link[=long_cluster_data]{long_cluster_data()}},
\code{\link[=long_gaussian_data]{long_gaussian_data()}},
\code{\link[=ortho_curve]{ortho_curve()}},
\code{\link[=random_circle_cluster_data]{random_circle_cluster_data()}},
\code{\link[=random_circle_data]{random_circle_data()}},
\code{\link[=random_jump]{random_jump()}},
\code{\link[=random_walk]{random_walk()}},
\code{\link[=simplex_data]{simplex_data()}},
\code{\link[=subset_clusters_data]{subset_clusters_data()}},
\code{\link[=three_clusters_data]{three_clusters_data()}},
\code{\link[=trefoil_data]{trefoil_data()}},
\code{\link[=two_clusters_data]{two_clusters_data()}},
\code{\link[=two_different_clusters_data]{two_different_clusters_data()}},
\code{\link[=unlink_data]{unlink_data()}}
}
\concept{distill functions}
