% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misread-tsne.R
\name{grid_data}
\alias{grid_data}
\title{2D Grid}
\usage{
grid_data(n = 10)
}
\arguments{
\item{n}{Number of points per side.}
}
\value{
Data frame with coordinates in the \code{x} and \code{y} columns,
 and a per-point color in the \code{color} column.
}
\description{
2D grid data from "How to Use t-SNE Effectively".
}
\details{
Creates a two-dimensional square grid with equal spacing between points.
Point color goes from dark to light with increasing \code{x} coordinate and
from green to blue with increasing \code{y}.
}
\examples{
df <- grid_data(n = 10)
}
\references{
\url{http://distill.pub/2016/misread-tsne/}
}
\seealso{
Other distill functions:
\code{\link[=circle_data]{circle_data()}},
\code{\link[=cube_data]{cube_data()}},
\code{\link[=gaussian_data]{gaussian_data()}},
\code{\link[=link_data]{link_data()}},
\code{\link[=long_cluster_data]{long_cluster_data()}},
\code{\link[=long_gaussian_data]{long_gaussian_data()}},
\code{\link[=ortho_curve]{ortho_curve()}},
\code{\link[=random_circle_cluster_data]{random_circle_cluster_data()}},
\code{\link[=random_circle_data]{random_circle_data()}},
\code{\link[=random_jump]{random_jump()}},
\code{\link[=random_walk]{random_walk()}},
\code{\link[=simplex_data]{simplex_data()}},
\code{\link[=subset_clusters_data]{subset_clusters_data()}},
\code{\link[=three_clusters_data]{three_clusters_data()}},
\code{\link[=trefoil_data]{trefoil_data()}},
\code{\link[=two_clusters_data]{two_clusters_data()}},
\code{\link[=two_different_clusters_data]{two_different_clusters_data()}},
\code{\link[=unlink_data]{unlink_data()}}
}
\concept{distill functions}
