% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{helix}
\alias{helix}
\title{Toroidal Helix Data Set}
\usage{
helix(n = 1000, rmajor = 2, rminor = 1, nwinds = 8)
}
\arguments{
\item{n}{Number of points to create.}

\item{rmajor}{Major radius.}

\item{rminor}{Minor radius.}

\item{nwinds}{Number of winds the helix makes.}
}
\value{
Data frame with \code{x}, \code{y}, \code{z} columns containing the
 coordinates of the points and \code{color} the RGB color.
}
\description{
Simulation data in the shape of a 3D toroidal helix.
}
\details{
Creates a series of points sampled from a 3D helix with the ends joined
to each other.

Unlike \code{\link{ball}} and \code{\link{sphere}}, this data set is not
randomly sampled.

Points are colored based on their distances from the origin.
}
\examples{
\dontrun{
 helix1000 <- helix(n = 1000)
 # Coloring should be obvious with a 2D plot of the cross section:
 plot(helix1000$x, helix1000$y, col = helix1000$color, pch = 20)
}
}
\references{
A dataset like this was used in:
Lee, J. A., Peluffo-Ordo'nez, D. H., & Verleysen, M. (2015).
Multi-scale similarities in stochastic neighbour embedding: Reducing
dimensionality while preserving both local and global structure.
\emph{Neurocomputing}, \emph{169}, 246-261.
}
