% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misread-tsne.R
\name{link_data}
\alias{link_data}
\title{Two Linked Rings}
\usage{
link_data(n)
}
\arguments{
\item{n}{Number of points to sample.}
}
\value{
Data frame with coordinates in the \code{x}, \code{y}, \code{z}
columns, and color in the \code{color} column.
}
\description{
Points sampled from two linked circles, from "How to Use t-SNE
Effectively".
}
\details{
Creates a 3D dataset where points are uniformly sampled from two (2D)
circles of unit radius, separated by 1 unit, and hence are overlapping.
The two circles are not co-planar. Points are colored depending from which
circle they are sampled from.

For the unlinked version, see \code{\link[=unlink_data]{unlink_data()}}.
}
\examples{
df <- link_data(n = 50)
}
\references{
\url{http://distill.pub/2016/misread-tsne/}
}
\seealso{
Other distill functions:
\code{\link[=circle_data]{circle_data()}},
\code{\link[=cube_data]{cube_data()}},
\code{\link[=gaussian_data]{gaussian_data()}},
\code{\link[=grid_data]{grid_data()}},
\code{\link[=long_cluster_data]{long_cluster_data()}},
\code{\link[=long_gaussian_data]{long_gaussian_data()}},
\code{\link[=ortho_curve]{ortho_curve()}},
\code{\link[=random_circle_cluster_data]{random_circle_cluster_data()}},
\code{\link[=random_circle_data]{random_circle_data()}},
\code{\link[=random_jump]{random_jump()}},
\code{\link[=random_walk]{random_walk()}},
\code{\link[=simplex_data]{simplex_data()}},
\code{\link[=subset_clusters_data]{subset_clusters_data()}},
\code{\link[=three_clusters_data]{three_clusters_data()}},
\code{\link[=trefoil_data]{trefoil_data()}},
\code{\link[=two_clusters_data]{two_clusters_data()}},
\code{\link[=two_different_clusters_data]{two_different_clusters_data()}},
\code{\link[=unlink_data]{unlink_data()}}
}
\concept{distill functions}
