% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misread_tsne.R
\name{long_cluster_data}
\alias{long_cluster_data}
\title{Long Cluster Data}
\usage{
long_cluster_data(n)
}
\arguments{
\item{n}{Number of points to sample per cluster.}
}
\value{
Data frame with coordinates in the \code{x}, \code{y} columns, and
color in the \code{color} column.
}
\description{
Two long, linear clusters in 2D from "How to Use t-SNE Effectively".
}
\details{
Creates a 2D dataset where points are sampled from two long, linear,
closely-separated clusters. Points are colored depending from which cluster
they are sampled from.
}
\examples{
df <- long_cluster_data(n = 50)
}
\references{
\url{http://distill.pub/2016/misread-tsne/}
}
\seealso{
Other distill functions: 
\code{\link{circle_data}()},
\code{\link{cube_data}()},
\code{\link{gaussian_data}()},
\code{\link{grid_data}()},
\code{\link{link_data}()},
\code{\link{long_gaussian_data}()},
\code{\link{ortho_curve}()},
\code{\link{random_circle_cluster_data}()},
\code{\link{random_circle_data}()},
\code{\link{random_jump}()},
\code{\link{random_walk}()},
\code{\link{simplex_data}()},
\code{\link{subset_clusters_data}()},
\code{\link{three_clusters_data}()},
\code{\link{trefoil_data}()},
\code{\link{two_clusters_data}()},
\code{\link{two_different_clusters_data}()},
\code{\link{unlink_data}()}
}
\concept{distill functions}
