% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olivetti.R
\name{olivetti_faces}
\alias{olivetti_faces}
\title{Olivetti Faces dataset}
\format{
A data frame with 400 rows and 4097 variables.
}
\usage{
olivetti_faces()
}
\value{
The Olivetti Faces dataset as a data frame.
}
\description{
SNE benchmarking data.
}
\details{
Returns the Olivetti Faces dataset in a data frame reformatted to have one
face per row, rather than column. This is a series of 400 images (with
dimensions 64 x 64) of 40 individuals' faces, with ten different poses per
person.

The variables are as follows:
\itemize{
\item \code{px1}, \code{px2}, \code{px3} ... \code{px4096} 8-bit grayscale
pixel values (0-255). The pixel index starts at the top right of the image
(\code{px1}) and are then stored column-wise.
\item \code{Label}: A factor with levels \code{1} through \code{40} indicating the person.
}

Each row has a name with the format \code{face_pose}, where \code{face} is the index
of the face, and \code{pose} is the index of the pose, e.g. the row with name
\verb{20_10} is the tenth pose of the twentieth face.
}
\note{
Requires the
\href{https://cran.r-project.org/package=RnavGraphImageData}{RnavGraphImageData}
package to be installed.
}
\seealso{
\itemize{
\item Sam Roweis' dataset web page: \url{http://www.cs.nyu.edu/~roweis/data.html}.
\item Each row can be visualized as an image using \code{\link[=show_olivetti_face]{show_olivetti_face()}}.
}
}
