% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misread-tsne.R
\name{ortho_curve}
\alias{ortho_curve}
\title{Mutually Orthogonal Step Data}
\usage{
ortho_curve(n)
}
\arguments{
\item{n}{Number of points to sample.}
}
\value{
Data frame with coordinates in the \code{X1}, \code{X2} ...
 \code{Xn} columns, and color in the \code{color} column.
}
\description{
Points related by mutually orthogonal steps, from "How to Use t-SNE
Effectively".
}
\details{
Creates N points from an N-dimensional space, where each point is related
by a mutually orthogonal step. Coordinates from a triangular matrix, where
the coordinates in the triangle are one, and are zero everywhere else.
Points are linearly mapped from their row id to a rainbow color scheme.
}
\examples{
df <- ortho_curve(n = 50)
}
\references{
\url{http://distill.pub/2016/misread-tsne/}
}
\seealso{
Other distill functions:
\code{\link[=circle_data]{circle_data()}},
\code{\link[=cube_data]{cube_data()}},
\code{\link[=gaussian_data]{gaussian_data()}},
\code{\link[=grid_data]{grid_data()}},
\code{\link[=link_data]{link_data()}},
\code{\link[=long_cluster_data]{long_cluster_data()}},
\code{\link[=long_gaussian_data]{long_gaussian_data()}},
\code{\link[=random_circle_cluster_data]{random_circle_cluster_data()}},
\code{\link[=random_circle_data]{random_circle_data()}},
\code{\link[=random_jump]{random_jump()}},
\code{\link[=random_walk]{random_walk()}},
\code{\link[=simplex_data]{simplex_data()}},
\code{\link[=subset_clusters_data]{subset_clusters_data()}},
\code{\link[=three_clusters_data]{three_clusters_data()}},
\code{\link[=trefoil_data]{trefoil_data()}},
\code{\link[=two_clusters_data]{two_clusters_data()}},
\code{\link[=two_different_clusters_data]{two_different_clusters_data()}},
\code{\link[=unlink_data]{unlink_data()}}
}
\concept{distill functions}
