% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misread_tsne.R
\name{random_circle_cluster_data}
\alias{random_circle_cluster_data}
\title{Fuzzy 2D Circle}
\usage{
random_circle_cluster_data(n)
}
\arguments{
\item{n}{Number of points to define the circle. These are then jittered
as described in 'Details', so the total number of observations in the
data frame will be \code{20 * n}.}
}
\value{
Data frame with x and y coordinates in the columns \code{x} and
 \code{y} respectively, and color in the \code{color} column.
}
\description{
Clusters arranged in a circle simulation data function from "How to Use t-SNE
Effectively".
}
\details{
Creates a dataset where the points are located on the circumference of a
circle of unit radius via \code{\link{circle_data}}, where the
angle subtended by the point at the center of the circle has been uniformly
sampled.

Each point is then replicated 20 times, with a slight perturbation to the
original coordinate (using a normal distribution with a standard deviation
of 0.1). Only the twenty jittered points are added to the data frame. Each
point has a color linearly mapped from the angle of the original unjittered
point to a rainbow color scheme.
}
\examples{
df <- random_circle_data(n = 50)
}
\references{
\url{http://distill.pub/2016/misread-tsne/}
}
\seealso{
Other distill.functions: \code{\link{circle_data}},
  \code{\link{cube_data}}, \code{\link{gaussian_data}},
  \code{\link{grid_data}}, \code{\link{link_data}},
  \code{\link{long_cluster_data}},
  \code{\link{long_gaussian_data}},
  \code{\link{ortho_curve}},
  \code{\link{random_circle_data}},
  \code{\link{random_jump}}, \code{\link{random_walk}},
  \code{\link{simplex_data}},
  \code{\link{subset_clusters_data}},
  \code{\link{three_clusters_data}},
  \code{\link{trefoil_data}},
  \code{\link{two_clusters_data}},
  \code{\link{two_different_clusters_data}},
  \code{\link{unlink_data}}
}

