% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misread-tsne.R
\name{random_jump}
\alias{random_jump}
\title{Random Jump Data}
\usage{
random_jump(n, dim)
}
\arguments{
\item{n}{Number of points to sample.}

\item{dim}{Number of dimensions of the dataset.}
}
\value{
Data frame with coordinates in the \code{X1}, \code{X2} ...
\code{Xdim} columns, and color in the \code{color} column.
}
\description{
Simulation data from a random jump process, from "How to Use t-SNE
Effectively".
}
\details{
Creates a dataset similar to \code{\link[=random_walk]{random_walk()}} but with
additional gaussian noise added at each step. Points are linearly mapped from
their position in the walk to a rainbow color scheme.
}
\examples{
df <- random_jump(n = 100, dim = 100)
}
\references{
\url{http://distill.pub/2016/misread-tsne/}
}
\seealso{
Other distill functions:
\code{\link[=circle_data]{circle_data()}},
\code{\link[=cube_data]{cube_data()}},
\code{\link[=gaussian_data]{gaussian_data()}},
\code{\link[=grid_data]{grid_data()}},
\code{\link[=link_data]{link_data()}},
\code{\link[=long_cluster_data]{long_cluster_data()}},
\code{\link[=long_gaussian_data]{long_gaussian_data()}},
\code{\link[=ortho_curve]{ortho_curve()}},
\code{\link[=random_circle_cluster_data]{random_circle_cluster_data()}},
\code{\link[=random_circle_data]{random_circle_data()}},
\code{\link[=random_walk]{random_walk()}},
\code{\link[=simplex_data]{simplex_data()}},
\code{\link[=subset_clusters_data]{subset_clusters_data()}},
\code{\link[=three_clusters_data]{three_clusters_data()}},
\code{\link[=trefoil_data]{trefoil_data()}},
\code{\link[=two_clusters_data]{two_clusters_data()}},
\code{\link[=two_different_clusters_data]{two_different_clusters_data()}},
\code{\link[=unlink_data]{unlink_data()}}
}
\concept{distill functions}
