% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{s_curve}
\alias{s_curve}
\title{S-curve data set.}
\usage{
s_curve(n_samples = 100, noise = 0)
}
\arguments{
\item{n_samples}{The number of points to create.}

\item{noise}{Add random noise normally-distributed with mean 0 and standard
deviation \code{noise}.}
}
\value{
Data frame with \code{x}, \code{y}, \code{z} columns containing the
 coordinates of the points and \code{color} the RGB color.
}
\description{
Simulation data randomly sampled from an S-shaped curve. Translated from the
\href{https://scikit-learn.org/stable/index.html}{scikit-learn} Pythom
function \code{sklearn.datasets.make_s_curve}.
}
\details{
Creates a series of points sampled from an S-shaped curve in 3D, with
optional normally-distributed noise. The S shape is oriented such that you
should be able to see it if you plot the X and Z columns.

Points are colored based on their distance along the curve.
}
\references{
Buitinck, L., Louppe, G., Blondel, M., Pedregosa, F., Mueller, A.,
Grisel, O., ... & Varoquaux, G. (2013).
API design for machine learning software: experiences from the scikit-learn
project.
\emph{arXiv preprint} \emph{arXiv:1309.0238}.
}
