% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coil.R
\name{show_coil_object}
\alias{show_coil_object}
\title{Visualize a COIL object image}
\usage{
show_coil_object(df, object, pose)
}
\arguments{
\item{df}{Data frame returned by \code{\link[=download_coil20]{download_coil20()}} or \code{\link[=download_coil100]{download_coil100()}}.
Matrix-list results returned with \code{as = "matrix"} are also supported.}

\item{object}{Object id to display. COIL-20 contains objects 1 to 20 and
COIL-100 contains objects 1 to 100.}

\item{pose}{For COIL-20, the pose id to display, from 0 to 71. For COIL-100,
the viewing angle, from 0 to 355 in five-degree increments.}
}
\description{
Display a COIL-20 or COIL-100 object pose.
}
\examples{
\dontrun{
coil20 <- download_coil20()
show_coil_object(coil20, object = 5, pose = 4)
}
}
