% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isomap.R
\name{show_isomap_face}
\alias{show_isomap_face}
\title{Visualize Isomap face.}
\usage{
show_isomap_face(df, n, col = grDevices::gray(1:255/255), ...)
}
\arguments{
\item{df}{Data frame containing Isomap faces, as returned by
\code{\link[=download_isomap_faces]{download_isomap_faces()}}.}

\item{n}{Row index of the image to display.}

\item{col}{List of colors to use in the display.}

\item{...}{Other arguments passed onto the \code{\link[graphics:image]{graphics::image()}}
function.}
}
\description{
Display an image from the Isomap face-pose dataset.
}
