% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norb.R
\name{show_norb_object}
\alias{show_norb_object}
\title{Visualize NORB object.}
\usage{
show_norb_object(
  df,
  category = 0,
  instance = 0,
  elevation = 0,
  azimuth = 0,
  lighting = 0
)
}
\arguments{
\item{df}{Data frame containing NORB objects.}

\item{category}{The category of the object, an integer from 0 to 4.}

\item{instance}{The instance in the category, an integer from 0 to 9.}

\item{elevation}{The elevation of the camera, an integer from 0 to 8, 
representing elevations of 30 to 70 degrees from the horizontal, in
increments of 5 degrees.}

\item{azimuth}{The azimuth, an even integer from 0, 2, 4, ... to 34.
Multiply by ten to get the value in degrees.}

\item{lighting}{The lighting condition, represened by an integer in
the range 0-5.}
}
\description{
Display a NORB object as an image pair.
}
\details{
The NORB objects as stored as image pairs and displayed with the first image
in the pair on top, and the second image in the pair below it.
}
\examples{
\dontrun{
show_norb_object(norb, category = 2, instance = 6, elevation = 6, 
                 azimuth = 24, lighting = 2)
# Compare with example at https://github.com/ndrplz/small_norb
}
}
