% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misread_tsne.R
\name{simplex_data}
\alias{simplex_data}
\title{Simplex Data}
\usage{
simplex_data(n, noise = 0.5, color = "#003399")
}
\arguments{
\item{n}{Number of points (and hence dimensionality) of the simplex.}

\item{noise}{Degree by which to perturb the points from their original
locations. A normal distribution with the specified standard deviation is
applied to the non-zero coordinate.}

\item{color}{Color to apply to each point.}
}
\value{
Data frame with coordinates in the \code{X1}, \code{X2} ...
 \code{Xn} columns, and color in the \code{color} column.
}
\description{
Data in a rough simplex from "How to Use t-SNE Effectively".
}
\details{
Create a dataset of the specified number of points arranged in a rough
simplex, which also determines the dimensionality of the dataset. Each
point has a single non-zero coordinate in a different column to all the
other points.
}
\examples{
df <- simplex_data(n = 50)
}
\references{
\url{http://distill.pub/2016/misread-tsne/}
}
\seealso{
Other distill.functions: \code{\link{circle_data}},
  \code{\link{cube_data}}, \code{\link{gaussian_data}},
  \code{\link{grid_data}}, \code{\link{link_data}},
  \code{\link{long_cluster_data}},
  \code{\link{long_gaussian_data}},
  \code{\link{ortho_curve}},
  \code{\link{random_circle_cluster_data}},
  \code{\link{random_circle_data}},
  \code{\link{random_jump}}, \code{\link{random_walk}},
  \code{\link{subset_clusters_data}},
  \code{\link{three_clusters_data}},
  \code{\link{trefoil_data}},
  \code{\link{two_clusters_data}},
  \code{\link{two_different_clusters_data}},
  \code{\link{unlink_data}}
}

