% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sphere}
\alias{sphere}
\title{Sphere data set.}
\usage{
sphere(n = 1000)
}
\arguments{
\item{n}{The number of points to create.}
}
\value{
Data frame with \code{x}, \code{y}, \code{z} columns containing the
 coordinates of the points and \code{color} the RGB color.
}
\description{
Simulation data randomly sampled from the surface of a 3D sphere.
}
\details{
Creates a series of points sampled from a 3D spherical surface.

Points are colored based on the angle theta in their spherical coordinate
presentation. You can think of it as dividing the sphere surface like the
segments of an orange.
}
\references{
A dataset like this was used in:
Lee, J. A., Peluffo-Ordo'nez, D. H., & Verleysen, M. (2015).
Multi-scale similarities in stochastic neighbour embedding: Reducing
dimensionality while preserving both local and global structure.
\emph{Neurocomputing}, \emph{169}, 246-261.
}
