% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic-hierarchical.R
\name{synthetic_hierarchical_data}
\alias{synthetic_hierarchical_data}
\title{Synthetic Hierarchical Data Set}
\usage{
synthetic_hierarchical_data(n = 500, dim = 50)
}
\arguments{
\item{n}{Number of observations to sample from each micro cluster.}

\item{dim}{Dimension of the Gaussian observations and cluster centers.}
}
\value{
Data frame with coordinates in the \code{X1}, \code{X2} ... \code{Xdim} columns,
factor columns \code{macro_label}, \code{meso_label}, and \code{micro_label}, a \code{color}
column containing the meso-level color, and \code{macro_color}, \code{meso_color},
and \code{micro_color} columns for plotting at each hierarchy level.
}
\description{
Simulation data from a hierarchy of Gaussian clusters.
}
\details{
Creates the synthetic hierarchical data set described by Wang and co-workers
(2021): five macro clusters, each containing five meso clusters, each
containing five micro clusters. By default each micro cluster contains 500
observations in 50 dimensions, for 62,500 observations in total.

Macro cluster centers are sampled from a zero-centered multivariate normal
distribution with diagonal covariance 10000. Meso cluster centers are sampled
around their macro center with diagonal covariance 1000. Micro cluster
centers are sampled around their meso center with diagonal covariance 100.
Observations are sampled around their micro center with diagonal covariance
10.

This is based on a Python notebook which can be found at
\url{https://github.com/jlmelville/drnb/blob/d542c97b26d6f0f481d7551b130c41fcc0206c6b/notebooks/data-pipeline/synthetic-hierarchical.ipynb}
}
\examples{
\dontrun{
df <- synthetic_hierarchical_data()
plot(df$X1, df$X2, col = df$color, pch = 20)
}
}
\references{
Wang, Y., Huang, H., Rudin, C., & Shaposhnik, Y. (2021).
Understanding how dimension reduction tools work: an empirical approach to
deciphering t-SNE, UMAP, TriMAP, and PaCMAP for data visualization.
\emph{J Mach. Learn. Res}, \emph{22}, 1-73.
}
\seealso{
The \href{https://github.com/hyhuang00/scRNA-DR2020/}{gitub repo} for
the paper \href{https://doi.org/10.1038/s42003-022-03628-x}{"Towards a comprehensive evaluation of dimension reduction methods for transcriptomic data visualization"},
which contains a numpy-formatted version of data generated under a similar
distribution, under the name \code{hierarchical_threelayer_dataset}.
}
