% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{taspheres}
\alias{taspheres}
\title{High Dimensional Spheres Dataset}
\usage{
taspheres(n_samples = 500, d = 100, n_spheres = 11, r = 5)
}
\arguments{
\item{n_samples}{Number of points to sample from each of the \code{n_spheres}
d-spheres. The larger d-sphere has \code{10 * n_samples} points.}

\item{d}{The dimensionality of each sphere. The returned dataframe will have
the \code{d + 1} dimensions of the Euclidean space in which the sphere is 
embedded.}

\item{n_spheres}{Number of spheres to return. There will be 
\code{n_spheres - 1} small spheres and 1 larger sphere.}

\item{r}{The radius of each of the smaller spheres. The larger sphere has 
radius \code{5 * r}.}
}
\value{
Data frame with \code{d + 1} numerical columns containing the
  coordinates of the d-spheres and a \code{"label"} factor column giving the
  identity of each d-sphere: levels \code{0 .. n_spheres - 2} are the smaller
  d-spheres. Level \code{n_spheres - 1} is the label for the big d-sphere.
}
\description{
Creates a dataframe consisting of samples from the d-spheres of radius
\code{r} enclosed within a larger d-sphere of radius \code{5 * r}.
}
\details{
This dataset was used by Moor and co-workers in their "Topological
Autoencoders" paper and this function is based on the Python code in the 
github repo for the paper.
}
\references{
Moor, M., Horn, M., Rieck, B., & Borgwardt, K. (2020).
Topological Autoencoders.
In \emph{Proceedings of the 37th International Conference on Machine Learning (ICML)}
(pp. 7045–7054). PMLR.

\url{https://michaelmoor.ml/blog/topoae/main/}

\url{https://github.com/BorgwardtLab/topological-autoencoders}
}
