% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misread_tsne.R
\name{three_clusters_data}
\alias{three_clusters_data}
\title{Three Gaussian Clusters With Equal Standard Deviations}
\usage{
three_clusters_data(n, dim = 50)
}
\arguments{
\item{n}{Number of points per gaussian.}

\item{dim}{Dimension of the gaussians.}
}
\value{
Data frame with coordinates in the \code{X1}, \code{X2} ...
 \code{Xdim} columns, and color in the \code{color} column.
}
\description{
Three gaussian clusters with equal size and bandwidth from "How to Use t-SNE
Effectively".
}
\details{
Creates a dataset consisting of three symmetric gaussian distributions with
equal number of points and standard deviation 1. The clusters are arranged
in a line, with the second cluster 10 units from the first, and the third
cluster 50 units from the first. Points are colored depending on which
cluster they belong to.
}
\examples{
df <- three_clusters_data(n = 50, dim = 2)
}
\references{
\url{http://distill.pub/2016/misread-tsne/}
}
\seealso{
Other distill functions: 
\code{\link{circle_data}()},
\code{\link{cube_data}()},
\code{\link{gaussian_data}()},
\code{\link{grid_data}()},
\code{\link{link_data}()},
\code{\link{long_cluster_data}()},
\code{\link{long_gaussian_data}()},
\code{\link{ortho_curve}()},
\code{\link{random_circle_cluster_data}()},
\code{\link{random_circle_data}()},
\code{\link{random_jump}()},
\code{\link{random_walk}()},
\code{\link{simplex_data}()},
\code{\link{subset_clusters_data}()},
\code{\link{trefoil_data}()},
\code{\link{two_clusters_data}()},
\code{\link{two_different_clusters_data}()},
\code{\link{unlink_data}()}
}
\concept{distill functions}
