% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misread_tsne.R
\name{two_clusters_data}
\alias{two_clusters_data}
\title{Two Equal Size Clusters}
\usage{
two_clusters_data(n, dim = 50)
}
\arguments{
\item{n}{Number of points per gaussian.}

\item{dim}{Dimension of the gaussians. You may pass a vector of length 2 to
create clusters of different dimensionalities, with the smaller cluster
having zeros in the extra dimensions.}
}
\value{
Data frame with coordinates in the \code{X1}, \code{X2} ...
 \code{Xdim} columns, and color in the \code{color} column.
}
\description{
Two gaussians with equal size and bandwidth, from "How to Use t-SNE
Effectively".
}
\details{
Creates a dataset consisting of two symmetric gaussian distributions with
equal number of points and standard deviation 1, separated by a distance
of 10 units. Points are colored depending on which cluster they belong to.
}
\examples{
df <- two_clusters_data(n = 50, dim = 2)
# two clusters with 10 members each, first 10 sampled from a 3D gaussian,
# second 10 are sampled from a 4D gaussian
df <- two_clusters_data(n = 10, dim = c(3, 4))
}
\references{
\url{http://distill.pub/2016/misread-tsne/}
}
\seealso{
Other distill functions: \code{\link{circle_data}},
  \code{\link{cube_data}}, \code{\link{gaussian_data}},
  \code{\link{grid_data}}, \code{\link{link_data}},
  \code{\link{long_cluster_data}},
  \code{\link{long_gaussian_data}},
  \code{\link{ortho_curve}},
  \code{\link{random_circle_cluster_data}},
  \code{\link{random_circle_data}},
  \code{\link{random_jump}}, \code{\link{random_walk}},
  \code{\link{simplex_data}},
  \code{\link{subset_clusters_data}},
  \code{\link{three_clusters_data}},
  \code{\link{trefoil_data}},
  \code{\link{two_different_clusters_data}},
  \code{\link{unlink_data}}
}
\concept{distill functions}
