% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misread_tsne.R
\name{unlink_data}
\alias{unlink_data}
\title{Two Unlinked Rings}
\usage{
unlink_data(n)
}
\arguments{
\item{n}{Number of points to sample.}
}
\value{
Data frame with coordinates in the \code{x}, \code{y}, \code{z}
columns, and color in the \code{color} column.
}
\description{
Points sampled from two unlinked circles, from "How to Use t-SNE
Effectively".
}
\details{
Creates a 3D dataset where points are uniformly sampled from two (2D)
circles of unit radius, separated by 3 units. The two circles are not
co-planar. Points are colored depending from which circle they are sampled
from.

For the linked version, see \code{\link{link_data}}.
}
\examples{
df <- unlink_data(n = 50)
}
\references{
\url{http://distill.pub/2016/misread-tsne/}
}
\seealso{
Other distill functions: 
\code{\link{circle_data}()},
\code{\link{cube_data}()},
\code{\link{gaussian_data}()},
\code{\link{grid_data}()},
\code{\link{link_data}()},
\code{\link{long_cluster_data}()},
\code{\link{long_gaussian_data}()},
\code{\link{ortho_curve}()},
\code{\link{random_circle_cluster_data}()},
\code{\link{random_circle_data}()},
\code{\link{random_jump}()},
\code{\link{random_walk}()},
\code{\link{simplex_data}()},
\code{\link{subset_clusters_data}()},
\code{\link{three_clusters_data}()},
\code{\link{trefoil_data}()},
\code{\link{two_clusters_data}()},
\code{\link{two_different_clusters_data}()}
}
\concept{distill functions}
