% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{ball}
\alias{ball}
\title{Ball Data Set}
\usage{
ball(n = 1000, rad = 1)
}
\arguments{
\item{n}{Number of points to create.}

\item{rad}{Radius of the ball.}
}
\value{
Data frame with \code{x}, \code{y}, \code{z} columns containing the
 coordinates of the points and \code{color} the RGB color.
}
\description{
Simulation data.
}
\details{
Creates a series of points sampled from a 3D spherical volume. Points are
colored based on the square of their distance from the origin.
}
\references{
A dataset like this was used in:
Lee, J. A., Peluffo-Ordo'nez, D. H., & Verleysen, M. (2015).
Multi-scale similarities in stochastic neighbour embedding: Reducing
dimensionality while preserving both local and global structure.
\emph{Neurocomputing}, \emph{169}, 246-261.
}

