% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norb.R
\name{download_norb_small}
\alias{download_norb_small}
\title{Download Small NORB}
\format{A data frame with 18,439 variables:

\describe{
\item{\code{c0px1}, \code{c0px2}, \code{c0px3} ... \code{c0px9216}}{Integer 
  pixel value, from 0 (white) to 255 (black) for the first image in the 
  pair}
\item{\code{c1px1}, \code{c1px2}, \code{c1px3} ... \code{c1px9216}}{Integer 
  pixel value, from 0 (white) to 255 (black) for the second image in the 
  pair}
\item{\code{Instance}}{The index of the toy in a particular category, 
  represented by a factor in the range 0-9. The training set consists of
  instances 0, 1, 2, 3 and 5, and the test set consists of 4, 6, 7, 8 and 9.}
\item{\code{Elevation}}{The elevation of the camera represented by a 
  factor in the range 0-8. These represent elevations of 30 to 70 degrees
  from the horizontal, in increments of 5 degrees.}
\item{\code{Azimuth}}{The azimuth, represented by a factor in the range 
  0, 2, 4 .. 34. Multiply by ten to get the value in degrees.}
\item{\code{Lighting}}{The lighting condition, represened by a factor in
  the range 0-5.}
\item{\code{Label}}{The toy category, represented by a factor in the 
  range 0-4.}
\item{\code{Split}}{Whether the toy in is in the \code{training} or 
  \code{testing} set, represented by a factor}
\item{\code{Description}}{The name of the toy category associated with 
  \code{Label}, represented by a factor.}
}

The pixel features are organized row-wise from the top left of each image. 
The \code{Label} levels correspond to:
\describe{
  \item{\code{0}}{Four-legged animal}
  \item{\code{1}}{Human figure}
  \item{\code{2}}{Airplane}
  \item{\code{3}}{Truck}
  \item{\code{4}}{Car}
}

There are 48,600 items in the data set. The first 24,300 are the training
set, and the remaining 24,300 are the testing set, but you can also use
the \code{Split} column to determine which split a given row is in.

Items in the dataset can be visualized with the
\code{\link{show_norb_object}} function.

For more information see
\url{https://cs.nyu.edu/~ylclab/data/norb-v1.0-small/}.}
\usage{

  download_norb_small(base_url = "https://cs.nyu.edu/~ylclab/data/norb-v1.0-small/",
  verbose = FALSE)
}
\arguments{
\item{base_url}{Base URL that the files are located at.}

\item{verbose}{If \code{TRUE}, then download progress will be logged as a
message.}
}
\value{
Data frame containing Small NORB dataset.
}
\description{
Download Small NORB database of images of toys.
}
\details{
Downloads the image and label files for the training and test datasets and
converts them to a data frame.

The Small NORB dataset contains images of 50 toys. The toys are divided into
five categories (animal, human, airplane, truck, car) with ten examples per
category. Each object was then images under 6 different lighting conditions,
9 elevations and 18 different azimuths, so there are 972 images per toy. The
process was then repeated with a different camera, so there are actually 972
* 2 = 1944 images per toy. This dataset stores each pair of images for a
given toy, lighting, elevation and azimuth as a single row. Each image is 96
* 96 pixels, so the first 9,216 columns contain the pixels of the first
image, and the second 9,216 (\code{9217:18432}) columns contain the pixels of
the second image. The other information (lighting and so on) are also stored
as factors.
}
\examples{
\dontrun{
# download the data set
norb <- download_norb_small(verbose = TRUE)

# first 24,300 instances are the training set
norb_train <- head(norb, 24300)
# the remaining 24,300 are the test set
norb_test <- tail(norb, 24300)

# Or equivalently
norb_train <- norb[norb$Split == "training", ]
norb_test <- norb[norb$Split == "testing", ]

identical(norb_train, norb_train2) # TRUE
identical(norb_test, norb_test2) # also TRUE

# PCA on 1000 examples
norb_r1000 <- norb[sample(nrow(norb), 1000), ]
pca <- prcomp(norb_r1000[, 1:(96 * 96 * 2)], retx = TRUE, rank. = 2)
# plot the scores of the first two components
plot(pca$x[, 1:2], type = 'n')
text(pca$x[, 1:2], labels = norb_r1000$Label,
     col = rainbow(length(levels(norb_r1000$Label)))[norb_r1000$Label])
}
}
\references{
The Small NORB Dataset, v1.0
\url{https://cs.nyu.edu/~ylclab/data/norb-v1.0-small/}

LeCun, Y., Huang, F. J., & Bottou, L. (2004, June). 
Learning methods for generic object recognition with invariance to pose and lighting. 
In \emph{IEEE Computer Society Conference on Computer Vision and Pattern Recognition (CVPR) 2004} 
(pp. 97-104).
IEEE.
\url{http://doi.ieeecomputersociety.org/10.1109/CVPR.2004.144}
}
