% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misread_tsne.R
\name{long_gaussian_data}
\alias{long_gaussian_data}
\title{Elongated Gaussian Ellipsoid}
\usage{
long_gaussian_data(n, dim, color = NULL)
}
\arguments{
\item{n}{Number of points.}

\item{dim}{Dimension of the gaussian.}

\item{color}{Color to assign to the data points.}
}
\value{
Data frame with coordinates in the \code{X1}, \code{X2} ...
 \code{Xdim} columns, and color in the \code{color} column.
}
\description{
Elongated gaussian data from "How to Use t-SNE Effectively".
}
\details{
Creates a randomly-sampled gaussian dataset of the specified
dimension, with standard deviation following the harmonic series: the first
column has standard deviation 1, the second 1/2, the third 1/3 and so on.
}
\examples{
df <- long_gaussian_data(n = 50, dim = 2)
}
\references{
\url{http://distill.pub/2016/misread-tsne/}
}
\seealso{
Other distill functions: \code{\link{circle_data}},
  \code{\link{cube_data}}, \code{\link{gaussian_data}},
  \code{\link{grid_data}}, \code{\link{link_data}},
  \code{\link{long_cluster_data}},
  \code{\link{ortho_curve}},
  \code{\link{random_circle_cluster_data}},
  \code{\link{random_circle_data}},
  \code{\link{random_jump}}, \code{\link{random_walk}},
  \code{\link{simplex_data}},
  \code{\link{subset_clusters_data}},
  \code{\link{three_clusters_data}},
  \code{\link{trefoil_data}},
  \code{\link{two_clusters_data}},
  \code{\link{two_different_clusters_data}},
  \code{\link{unlink_data}}
}
