% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnist.R
\name{show_mnist_digit}
\alias{show_mnist_digit}
\title{Visualize MNIST digit.}
\usage{
show_mnist_digit(df, n, col = grDevices::gray(1:255/255), ...)
}
\arguments{
\item{df}{Data frame containing MNIST digits.}

\item{n}{Row index of the digit to display.}

\item{col}{List of colors to use in the display.}

\item{...}{Other arguments passed onto the \code{\link[graphics]{image}}
function.}
}
\description{
Display an MNIST digit as an image the hand-written digit represented by the
nth row in a data frame.
}
\note{
Originally based on a function by Brendan O'Connor, which can be found
at \url{https://gist.github.com/brendano/39760}.
}
\examples{
\dontrun{
# show the fifth digit
mnist <- download_mnist()
show_mnist_digit(mnist, 5)
}
}
