% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{swiss_roll}
\alias{swiss_roll}
\title{Swiss Roll Data Set}
\usage{
swiss_roll(n = 1000)
}
\arguments{
\item{n}{Number of points to create.}
}
\value{
Data frame with \code{x}, \code{y}, \code{z} columns containing the
 coordinates of the points and \code{color} the RGB color.
}
\description{
Simulation data randomly sampled from a 2D plane curled up into 3 dimensions.
}
\details{
Creates a series of points randomly sampled from a swiss roll-shaped
manifold: a two-dimensional plane which has been rolled up into a spiral
shape. Or just look at a swiss roll.

The formula for sampling the x, y and z coordinates used in this dataset is
from that given in the Stochastic Proximity Embedding paper by Agrafiotis
and Xu:
\deqn{x = \phi cos\phi, y = \phi sin \phi, z}
{x = phi * cos(phi), y = phi * sin(phi), z}

where \eqn{\phi}{phi} and z are random numbers in the intervals [5, 13]
and [0, 10], respectively.

Points are colored based on the value of phi. If you unrolled the manifold
into a flat sheet, you would see the color change linearly in the direction
you unrolled it. Or just plot the x-y cross section (see the examples).
}
\examples{
\dontrun{
 swiss1000 <- swiss_roll(n = 1000)
 # Coloring should be obvious with a 2D plot of the cross section:
 plot(swiss1000$x, swiss1000$y, col = swiss1000$color, pch = 20)
}
}
\references{
Agrafiotis, D. K., & Xu, H. (2002).
A self-organizing principle for learning nonlinear manifolds.
\emph{Proceedings of the National Academy of Sciences}, \emph{99}(25), 15869-15872.
}

